gpt4o-realtime-transcribe/speech-to-text/requirements.txt
pyaudio==0.2.14
websocket-client==1.7.0
python-dotenv==1.0.0